+++ src/core/API/MolpherMolImpl.hpp
@@ -52,7 +52,9 @@ public:
 	std::vector<std::shared_ptr<MolpherAtom>> getNeighbors(int idx);
     std::unique_ptr<MolpherMolImpl> copy() const;
 
-	std::string asMolBlock() const;
+	std::string asMolBlock(bool include_locks = false) const;
+	RDKit::RWMol* asRDMol(bool include_locks = false) const;
+	const std::vector<std::shared_ptr<MolpherAtom>>& getAtoms() const;
 };
 
 #endif	/* MOLPHERMOLIMPL_HPP */