+++ src/python/molpher/core/morphing/__init__.py
@@ -1,2 +1,3 @@
 from molpher.core.morphing.AtomLibrary import AtomLibrary
 from molpher.core.morphing.Molpher import Molpher
+from molpher.core.morphing.MorphCollector import MorphCollector