+++ src/python/molpher/swig_wrappers/core.py
@@ -1589,8 +1589,8 @@ class MolpherMol(_object):
     def copy(self):
         return _core.MolpherMol_copy(self)
 
-    def asMolBlock(self):
-        return _core.MolpherMol_asMolBlock(self)
+    def asMolBlock(self, include_locks=False):
+        return _core.MolpherMol_asMolBlock(self, include_locks)
 
     def lockAtom(self, idx, mask):
         return _core.MolpherMol_lockAtom(self, idx, mask)